Remove UHD dependencies and unsupported SDRs Remove UHD build and dependencies from Dockerfile/CMakeLists.txt Fix SDRplay API version to match actual files (3.15.2) Remove USRP, HackRF, Pluto source files from build Update Capture class to support only RSPDuo Fix cmake duplicate argument syntax error
